Fix duplicate active IP usage for same ipaddress
Sometimes we might miss the end event for a port and so we don't
set the end on the ip usage. Whenever we create a new ipusage we
should check for this case and set the old active one to ended.

We choose 1 second before the new port is created, we have no way of
telling when exactly this is so this is a best guess.

Change-Id: I173377607f84e52b85d2b5c8f822bef918d75343